invoice-automation/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pymongo==4.11.3
pytesseract==0.3.13
opencv-python-headless==4.11.0.86
pdf2image==1.17.0
pandas==2.2.3
scikit-learn==1.6.1
python-dotenv==1.1.0
werkzeug==3.1.3
openai==1.55.3
